Mouse callbacks for Frame. Based on ImGui

+++ src/shader/shaderBase.cpp
@@ -149,7 +149,7 @@ void Shader::disable() const {
     glUseProgram(0);
 }
 
-void Shader::setViewPort(const glm::vec2& leftBottom, const glm::vec2& size) {
+void Shader::setViewPort(const glm::ivec2& leftBottom, const glm::ivec2& size) {
     glViewport(leftBottom.x, leftBottom.y, size.x, size.y);
 }
 void Shader::attr(const Attribute& attribute, const void* data, GLsizei stride, GLuint offset) {
